Create package extension for CUDA
