+++ GeneralisedFilters/src/resamplers.jl
@@ -1,8 +1,6 @@
 using Random
 using Distributions
 
-using AcceleratedKernels: searchsortedfirst
-
 export Multinomial, Systematic, Stratified, Metropolis, Rejection
 export ESSResampler
 
@@ -43,6 +41,15 @@ function maybe_resample(
     return resample(rng, resampler, state, weights; ref_state, auxiliary_weights)
 end
 
+"""
+    _set_ref_index!(idxs)
+
+Set the first ancestor index to 1 for conditional SMC (preserving the reference trajectory).
+This is a separate function so it can be overloaded for GPU arrays (which require
+`CUDA.@allowscalar` for scalar indexing).
+"""
+_set_ref_index!(idxs) = (idxs[1] = 1; nothing)
+
 function resample(
     rng::AbstractRNG,
     resampler::AbstractResampler,
@@ -54,7 +61,7 @@ function resample(
 )
     idxs = sample_ancestors(rng, resampler, weights)
     if !isnothing(ref_state)
-        CUDA.@allowscalar idxs[1] = 1
+        _set_ref_index!(idxs)
     end
     return construct_new_state(state, idxs, auxiliary_weights)
 end
@@ -224,17 +231,6 @@ function sample_ancestors(
     return rand(rng, Distributions.Categorical(weights), n)
 end
 
-# Following Code 5 of Murray et. al (2015)
-function sample_ancestors(
-    rng::AbstractRNG, ::Multinomial, weights::CuVector{WT}, n::Int=length(weights)
-) where {WT}
-    W = cumsum(weights)
-    Wn = CUDA.@allowscalar W[n]
-    us = CUDA.rand(WT, n) * Wn
-    as = searchsortedfirst(W, us)
-    return as
-end
-
 struct Systematic <: AbstractResampler end
 
 function sample_ancestors(
@@ -261,46 +257,7 @@ function sample_ancestors(
     return a
 end
 
-function sample_ancestors(
-    rng::AbstractRNG, ::Systematic, weights::CuVector, n::Int=length(weights)
-)
-    offspring = sample_offspring(rng, Systematic(), weights, n)
-    return offspring_to_ancestors(offspring)
-end
-
-# Following Code 8 of Murray et. al (2015)
-function sample_offspring(
-    rng::AbstractRNG, ::Systematic, weights::CuVector{WT}, n::Int=length(weights)
-) where {WT}
-    W = cumsum(weights)
-    Wn = CUDA.@allowscalar W[n]
-    u0 = CUDA.@allowscalar rand(rng, WT)
-    r = n * W / Wn
-    offspring = min.(n, floor.(Int, r .+ u0))
-    return offspring
-end
-
 struct Stratified <: AbstractResampler end
-
-function sample_ancestors(
-    rng::AbstractRNG, ::Stratified, weights::CuVector, n::Int=length(weights)
-)
-    offspring = sample_offspring(rng, Stratified(), weights, n)
-    return offspring_to_ancestors(offspring)
-end
-
-# Following Code 7 of Murray et. al (2015)
-function sample_offspring(
-    rng::AbstractRNG, ::Stratified, weights::CuVector{WT}, n::Int=length(weights)
-) where {WT}
-    u = rand(rng, n)
-    W = cumsum(weights)
-    Wn = CUDA.@allowscalar W[n]
-    r = n * W / Wn
-    k = min.(n, floor.(Int, r .+ 1))
-    offspring = min.(n, floor.(Int, r .+ u[k]))
-    return offspring
-end
 
 ## SINGLE PRECISION STABLE ALGORITHMS ######################################################
 
@@ -363,60 +320,4 @@ function sample_ancestors(
     end
 
     return a
-end
-
-## ANCESTOR-OFFSPRING CONVERSION ###########################################################
-
-function _offspring_to_ancestors_kernel!(ancestors, offspring, N)
-    index = (blockIdx().x - 1) * blockDim().x + threadIdx().x
-    stride = blockDim().x * gridDim().x
-
-    @inbounds for i in index:stride:N
-        start = i == 1 ? 0 : offspring[i - 1]
-        finish = offspring[i]
-        for j in (start + 1):finish
-            ancestors[j] = i
-        end
-    end
-
-    return nothing
-end
-
-function offspring_to_ancestors(offspring::CuVector{<:Integer})
-    N = length(offspring)
-    ancestors = similar(offspring)
-
-    threads = 256
-    blocks = ceil(Int, N / threads)
-
-    @cuda threads = threads blocks = blocks _offspring_to_ancestors_kernel!(
-        ancestors, offspring, N
-    )
-
-    return ancestors
-end
-
-function _ancestors_to_offspring_kernel!(output, ancestors, N)
-    index = (blockIdx().x - 1) * blockDim().x + threadIdx().x
-    stride = blockDim().x * gridDim().x
-
-    @inbounds for i in index:stride:N
-        CUDA.@atomic output[ancestors[i]] += 1
-    end
-
-    return nothing
-end
-
-function ancestors_to_offspring(ancestors::CuVector{Int})
-    N = length(ancestors)
-    offspring = CUDA.zeros(Int, N)
-
-    threads = 256
-    blocks = ceil(Int, N / threads)
-
-    @cuda threads = threads blocks = blocks _ancestors_to_offspring_kernel!(
-        offspring, ancestors, N
-    )
-
-    return offspring
 end
